feat: add podium of tenacity and fix bugs related to podium (#1379)
Added:
• Parse to items.xml to read "transformOnUse".It should be used for simple transformations, which don't need to complex scripting.
• Podium of tenacity feature
• "/addbosskill" talkaction for add amount of kills (bosstiary) to certain player (or itself).

Fixed:
• Rotate from podium of vigor and renown.
• Size of "boss race id" to uint16_t, and now it uses the "raceid" from bestiary (it's the same).

I made some other small tweaks and improvements.

+++ src/io/io_bosstiary.hpp
@@ -42,8 +42,8 @@ class IOBosstiary {
 
 		void loadBoostedBoss();
 
-		void addBosstiaryMonster(uint32_t raceId, const std::string &name);
-		const phmap::btree_map<uint32_t, std::string> &getBosstiaryMap() const;
+		void addBosstiaryMonster(uint16_t raceId, const std::string &name);
+		const phmap::btree_map<uint16_t, std::string> &getBosstiaryMap() const;
 
 		const phmap::btree_map<BosstiaryRarity_t, std::vector<LevelInfo>> levelInfos = {
 			{ BosstiaryRarity_t::RARITY_BANE, { { 25, 5 }, { 100, 15 }, { 300, 30 } } },
@@ -69,7 +69,7 @@ class IOBosstiary {
 		IOBosstiary() = default;
 		~IOBosstiary() = default;
 
-		phmap::btree_map<uint32_t, std::string> bosstiaryMap;
+		phmap::btree_map<uint16_t, std::string> bosstiaryMap;
 		std::string boostedBoss;
 		uint16_t boostedBossId = 0;
 };
